Conflicts:
	CODE/CORE/DATA/STREAM/CORE_DATA_STREAM.h

+++ CODE/CORE/HELPERS/IDENTIFIER/CORE_HELPERS_UNIQUE_IDENTIFIER.cpp
@@ -18,7 +18,7 @@
 // http://stackoverflow.com/questions/8317508/hash-function-for-a-string
 
 XS_IMPLEMENT_INTERNAL_MEMORY_LAYOUT( CORE_HELPERS_UNIQUE_IDENTIFIER )
-    XS_DEFINE_ClassMemberArray( "Identifier", char, Identifier, (size_t) 32 )
+    XS_DEFINE_ClassMemberArray( "Identifier", char, Identifier, (int) 32 )
     XS_DEFINE_ClassMember( "CheckSum", unsigned int, CheckSum )
 XS_END_INTERNAL_MEMORY_LAYOUT
